cloud_app/services/auth-service/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
bcrypt==4.0.1
werkzeug==2.3.7
redis==4.6.0
email-validator==2.0.0
flask-limiter==3.5.0